requirements.txt
google.generativeai==0.8.3

tiktoken
faiss-cpu
langchain
PyPDF2
python-dotenv
streamlit